+++ assets/js/modules/analytics-4/components/settings/SettingsView.js
@@ -39,13 +39,9 @@ import DisplaySetting, {
 import Link from '../../../../components/Link';
 import VisuallyHidden from '../../../../components/VisuallyHidden';
 import { escapeURI } from '../../../../util/escape-uri';
-import { useFeature } from '../../../../hooks/useFeature';
-import { CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
 const { useSelect } = Data;
 
 export default function SettingsView() {
-	const iceEnabled = useFeature( 'conversionInfra' );
-
 	const accountID = useSelect( ( select ) =>
 		select( MODULES_ANALYTICS_4 ).getAccountID()
 	);
@@ -69,14 +65,6 @@ export default function SettingsView() {
 	const editDataStreamSettingsURL = useSelect( ( select ) =>
 		select( MODULES_ANALYTICS_4 ).getServiceEntityAccessURL()
 	);
-
-	const isConversionTrackingEnabled = useSelect( ( select ) => {
-		if ( ! iceEnabled ) {
-			return false;
-		}
-
-		return select( CORE_SITE ).isConversionTrackingEnabled();
-	} );
 
 	if ( ! propertyID || propertyID === PROPERTY_CREATE ) {
 		return null;
@@ -195,25 +183,6 @@ export default function SettingsView() {
 			</div>
 
 			<SettingsEnhancedMeasurementView />
-
-			{ iceEnabled && (
-				<div className="googlesitekit-settings-module__meta-item">
-					<h5 className="googlesitekit-settings-module__meta-item-type">
-						{ __(
-							'Enhanced Conversion Tracking',
-							'google-site-kit'
-						) }
-					</h5>
-					<p className="googlesitekit-settings-module__meta-item-data">
-						{ isConversionTrackingEnabled &&
-							__( 'Enabled', 'google-site-kit' ) }
-						{ isConversionTrackingEnabled === false &&
-							__( 'Disabled', 'google-site-kit' ) }
-						{ isConversionTrackingEnabled === undefined &&
-							BLANK_SPACE }
-					</p>
-				</div>
-			) }
 
 			<OptionalSettingsView />
 		</div>
